pyproject.toml
[tool.poetry]
name = "jinja2-org-template"
version = "0.1.0"
description = "Jinja2 template with org mode"
authors = ["Samuel Mesa <samuelmesa@linuxmail.org>"]
license = "GPL3"

[tool.poetry.dependencies]
python = "^3.10"
Jinja2 = "^3.1.1"
python-emacs = "^0.2.0"
matplotlib = "^3.5.1"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
